% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{addHeaderToSavedata}
\alias{addHeaderToSavedata}
\title{Add header to saved data}
\usage{
addHeaderToSavedata(outfile, directory = getwd())
}
\arguments{
\item{outfile}{The output file}

\item{director}{The current working directory by default}
}
\description{
Description
}
\examples{
# make me!!!
}
\keyword{internal}
