% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{checkSubmission}
\alias{checkSubmission}
\title{check on the status of submitted Mplus jobs on the cluster}
\usage{
checkSubmission(mplus_submission_df = NULL, quiet = FALSE)
}
\arguments{
\item{mplus_submission_df}{The data.frame returned by \code{submitModels} containing
jobs to check on}

\item{quiet}{If \code{TRUE}, do not print out the submission data.frame with current status}
}
\value{
invisibly, the \code{mplus_submission_df} with \verb{$status} amd \verb{$status_time} updated
}
\description{
check on the status of submitted Mplus jobs on the cluster
}
