% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.mplusObject}
\alias{coef.mplusObject}
\title{Extract coefficients from an mplusObject}
\usage{
\method{coef}{mplusObject}(object, ...)
}
\arguments{
\item{object}{An object of class mplusObject.}

\item{\dots}{Additional arguments passed to \code{coef}.}
}
\value{
Typically a data frame.
}
\description{
Method that calls \code{coef.mplus.model}.
See further documentation there.
}
\seealso{
Other Mplus.Formatting: \code{\link{coef.mplus.model}};
  \code{\link{extract,mplus.model-method}},
  \code{\link{extract,mplusObject-method}},
  \code{\link{extract.mplus.model}};
  \code{\link{print.MplusRstructure}};
  \code{\link{summary.mplusObject}}
}
\keyword{interface}

