% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{convert_to_filelist}
\alias{convert_to_filelist}
\title{helper function to crawl over the target location, determine if it is a file or folder,
then locate all .inp files, and convert them to absolute paths}
\usage{
convert_to_filelist(target, filefilter = NULL, recursive = FALSE)
}
\arguments{
\item{target}{a character vector where each element is a directory containing Mplus input files 
(.inp) to run OR a single .inp file to be run. Elements may be a full path, relative path,
or a filename within the working directory.}

\item{filefilter}{An optional PCRE expression for filtering input files of interest}

\item{recursive}{if TRUE, search for .inp files in subfolders of all elements of \code{target}}
}
\value{
A vector of .inp file locaitons
}
\description{
helper function to crawl over the target location, determine if it is a file or folder,
then locate all .inp files, and convert them to absolute paths
}
\keyword{internal}
