% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{createMixtures}
\alias{createMixtures}
\title{Create syntax for a batch of mixture models}
\usage{
createMixtures(classes = 1L, filename_stem = NULL, model_overall = NULL,
  model_class_specific = NULL, rdata = NULL, usevariables = NULL,
  SAVEDATA = "FILE IS {filename_stem}_{C}.dat;  SAVE = cprobabilities;", ...)
}
\arguments{
\item{classes}{A vector of integers, indicating which class solutions to
generate. Defaults to 1L. E.g., \code{classes = 1:6},
\code{classes = c(1:4, 6:8)}.}

\item{filename_stem}{Character. A stem for the automatically generated filenames of
the syntax and data files.}

\item{model_overall}{Character. Mplus syntax for the overall model (across
classes).}

\item{model_class_specific}{Character. Mplus syntax for the class-specific
model. The character string \dQuote{\{C\}} is substituted with the correct
class number, for example to make unique parameter labels for each class.}

\item{rdata}{Data.frame. An R dataset to be used for the model.}

\item{usevariables}{Character vector, specifying the names of variables in 
the rdata object which should be included in the Mplus data file and model.}

\item{SAVEDATA}{Character. Syntax for Mplus' savedata option. Highly
recommended when conducting mixture models. The default option will often be
adequate.}

\item{...}{Additional arguments, passed to \link{mplusObject}, such as syntax
for other Mplus options.}
}
\value{
None. Function is used for its side effects (generating syntax).
}
\description{
Dynamically creates syntax for a batch of mixture models, with intelligent
defaults. This function is a wrapper around \code{mplusObject}, and
additional arguments can be passed to this function using \code{...}.
In all arguments to \code{mplusObject}, a double space (\dQuote{  }) is
replaced with a newline character. This can be used to obtain nicely
formatted Mplus syntax.
In the arguments \code{model_class_specific} and \code{SAVEDATA}, the
character string \dQuote{\{C\}} is substituted with the correct class number.
The character string \dQuote{\{filename_stem\}} is substituted with the
filename stem, for example, to name savedata in line with the input files.
}
\examples{
createMixtures(classes = 1:4, filename_stem = "iris", rdata = iris)
}
\seealso{
\code{\link{mplusObject}}, \code{\link{mplusModeler}}
}
\author{
Caspar J. van Lissa
}
\keyword{mixture}
\keyword{models}
\keyword{mplus}
