% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{detectColumnNames}
\alias{detectColumnNames}
\title{Detect Column Names}
\usage{
detectColumnNames(filename, modelSection, sectionType = "model_results")
}
\arguments{
\item{filename}{The file name}

\item{modelSection}{The model section}

\item{sectionType}{A character string.  Defaults to \dQuote{model_results}.}
}
\value{
A list with elements for the directory, filename,
  and absolute path.
}
\description{
Helper function to detect model results columns.
}
\examples{
# make me!!!
}
\keyword{internal}
