% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{dhms_to_hours}
\alias{dhms_to_hours}
\title{helper function to convert a dhms string to the number of hours for combining similar job times}
\usage{
dhms_to_hours(str)
}
\arguments{
\item{str}{string in dhms format}
}
\value{
the number of hours represented by the dhms string
}
\description{
helper function to convert a dhms string to the number of hours for combining similar job times
}
\keyword{internal}
