% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{divideIntoFields}
\alias{divideIntoFields}
\title{Divide text into fields}
\usage{
divideIntoFields(section.text, required)
}
\arguments{
\item{section.text}{The section text}

\item{required}{Required sections}
}
\value{
Divided sections
}
\description{
Helper function to divide an input section into key-value pair list taken from mplus2lavaan
}
\examples{
# make me!!!
}
\keyword{internal}
