% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractAuxDis.R
\name{extractAux}
\alias{extractAux}
\title{Extract Auxiliary condition means and comparisons.}
\usage{
extractAux(outfiletext, filename)
}
\arguments{
\item{outfiletext}{}

\item{filename}{}
}
\value{
A data frame
}
\description{
To do: add details.
}
\keyword{internal}

