% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractDataSummary}
\alias{extractDataSummary}
\title{Function to extract the SUMMARY OF DATA section from Mplus outputs}
\usage{
extractDataSummary(outfiletext, filename)
}
\arguments{
\item{outfiletext}{The text of the output file}

\item{filename}{the name of the file containing textToScan. Used to make more intelligible warning messages.}
}
\description{
Function to extract the SUMMARY OF DATA section from Mplus outputs
}
\keyword{internal}
