% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEFAparameters.R
\name{extractEFAparameters}
\alias{extractEFAparameters}
\title{Extract the model parameters from an EFA Mplus model output}
\usage{
extractEFAparameters(outfiletext, filename)
}
\arguments{
\item{outfiletext}{character vector of Mplus output file being processed}

\item{filename}{name of the output file}
}
\value{
list of parsed EFA parameters
}
\description{
Extract the model parameters from an EFA Mplus model output
}
