% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndirect.R
\name{extractIndirect}
\alias{extractIndirect}
\title{Extract Indirect Effects output}
\usage{
extractIndirect(outfiletext, curfile)
}
\arguments{
\item{outfiletext}{a character vector containing the indirect effects output section returned by getSection}

\item{curFile}{the name of the current output file being parsed}
}
\value{
An mplus.indirect object (inherits list) containing $overall and $specific
}
\description{
This function parses the section titles TOTAL, TOTAL INDIRECT, SPECIFIC INDIRECT, AND DIRECT EFFECTS
It returns a list composed of $overall and $specific effects
}
\keyword{internal}
