% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndirect.R
\name{extractIndirect_section}
\alias{extractIndirect_section}
\title{Extract Indirect Effects output}
\usage{
extractIndirect_section(indirectSection, curfile, sectionType)
}
\arguments{
\item{indirectSection}{a character vector containing the indirect effects for a specific section (e.g., stdyx)}

\item{curfile}{the name of the current output file being parsed}
}
\value{
An mplus.indirect object (inherits list) containing $overall and $specific
}
\description{
This function parses a given indirect section
It returns a list composed of $overall and $specific effects
}
\keyword{internal}
