% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractInput_1file}
\alias{extractInput_1file}
\title{Extract and parse Mplus input file}
\usage{
extractInput_1file(outfiletext, filename)
}
\arguments{
\item{outfiletext}{The text of the output file}

\item{filename}{The filename}
}
\value{
The parsed input file
}
\description{
Function to extract and parse mplus input syntax from the output file
}
\examples{
# make me!!!
}
\keyword{internal}

