% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModIndices.R
\name{extractModIndices_1chunk}
\alias{extractModIndices_1chunk}
\title{Extract Modification Indices for One Chunk (Section)}
\usage{
extractModIndices_1chunk(chunk, columnNames, filename)
}
\arguments{
\item{chunk}{character vector of current subsection from which to extract mod indices}

\item{columnNames}{character vector of expected column (variable) names for section}

\item{filename}{name of Mplus output file being parsed}
}
\value{
A data frame
}
\description{
To do: add details.
}
\keyword{internal}

