% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModIndices.R
\name{extractModIndices_1file}
\alias{extractModIndices_1file}
\title{Extract Modification Indices for One File}
\usage{
extractModIndices_1file(outfiletext, filename)
}
\value{
A data frame
}
\description{
To do: add details.
}
\keyword{internal}

