% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParameters.R
\name{extractParameters_1chunk}
\alias{extractParameters_1chunk}
\title{Extract Parameters for One Chunk}
\usage{
extractParameters_1chunk(filename, thisChunk, columnNames, sectionName)
}
\arguments{
\item{filename}{name of Mplus output file being processed}

\item{thisChunk}{character vector of output chunk from which to extract parameter values}

\item{columnNames}{character vector of expected column names for output chunk}
}
\value{
A data frame (or matrix?)
}
\description{
Helper function for extractModelParameters. Used to parse each subsection of
output within a given file and given results section (e.g., stdyx section) There
will be many chunks if latent classes, multiple groups, multilevel features are used.
}
\keyword{internal}

