% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAuxDis.R
\name{extractR3step}
\alias{extractR3step}
\title{Extract output of R3STEP procedure}
\usage{
extractR3step(outfiletext, filename)
}
\arguments{
\item{outfiletext}{character vector of Mplus output file from which to extract the AUX section}

\item{filename}{filename of the Mplus output file being processed}
}
\value{
A list containing the parsed R3STEP sections
}
\description{
Extract output of R3STEP procedure
}
\keyword{internal}
