% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractResiduals}
\alias{extractResiduals}
\title{Extract residual matrices}
\usage{
extractResiduals(outfiletext, filename)
}
\arguments{
\item{outfiletext}{the text of the output file}

\item{filename}{The name of the file}
}
\value{
A list of the residual matrices
}
\description{
Function that extracts the residual matrices including standardized ones
}
\examples{
# make me!!!
}
\seealso{
\code{\link{matrixExtract}}
}
\keyword{internal}

