% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractResiduals_1section}
\alias{extractResiduals_1section}
\title{Helper subfunction to extract one section of OUTPUT: RESIDUALS
Can be called multiple times, as in the case of invariance testing}
\usage{
extractResiduals_1section(residSection)
}
\description{
Helper subfunction to extract one section of OUTPUT: RESIDUALS
Can be called multiple times, as in the case of invariance testing
}
