% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractSampstat}
\alias{extractSampstat}
\title{Helper function to extract the sample statistics from Mplus output
Depends on OUTPUT: SAMPSTAT}
\usage{
extractSampstat(outfiletext, filename)
}
\arguments{
\item{outfiletext}{The character vector containing all strings from output file}

\item{filename}{The current out file being parsed}
}
\description{
Helper function to extract the sample statistics from Mplus output
Depends on OUTPUT: SAMPSTAT
}
\keyword{internal}
