% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractSummaries_1plan}
\alias{extractSummaries_1plan}
\title{Worker function used in extractSummaries_1section}
\usage{
extractSummaries_1plan(arglist, sectionHeaders, sectionFields, textToParse,
  filename)
}
\arguments{
\item{arglist}{The argument list}

\item{sectionHeaders}{A character vector with headers for each section of interest}

\item{sectionFields}{is a list of data.frames where each data.frame specifies the fields to be extracted for that section}

\item{textToParse}{The text to parse}

\item{filename}{The filename}
}
\value{
A list
}
\description{
Worker function used in extractSummaries_1section
}
\examples{
# make me!!!
}
\keyword{internal}

