\name{extractTech10}
\alias{extractTech10}
\title{Extract Technical 10 matrix from Mplus}
\usage{
  extractTech10(outfiletext, filename)
}
\arguments{
  \item{outfiletext}{the text of the output file}

  \item{filename}{The name of the file}
}
\value{
  An empty list
}
\description{
  Function that extracts the Tech10 matrix
}
\examples{
# make me!!!
}
\seealso{
  \code{\link{matrixExtract}}
}
\keyword{internal}

