% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractTech15}
\alias{extractTech15}
\title{Extract Technical 15 from Mplus}
\usage{
extractTech15(outfiletext, filename)
}
\arguments{
\item{outfiletext}{the text of the output file}

\item{filename}{The name of the file}
}
\value{
A list of class \dQuote{mplus.tech15}
}
\description{
The TECH15 option is used in conjunction with TYPE=MIXTURE to request conditional probabilities
for the latent class variables.
}
\examples{
# make me!!!
}
\seealso{
\code{\link{matrixExtract}}
}
\keyword{internal}
