\name{extractTech3}
\alias{extractTech3}
\title{Extract Technical 3 matrix from Mplus}
\usage{
  extractTech3(outfiletext, savedata_info, filename)
}
\arguments{
  \item{outfiletext}{the text of the output file}

  \item{savedata_info}{Information on saved data}

  \item{filename}{The name of the file}
}
\value{
  A list of class \dQuote{mplus.tech3}
}
\description{
  Function that extracts the Tech3 matrix
}
\examples{
# make me!!!
}
\seealso{
  \code{\link{matrixExtract}}
}
\keyword{internal}

