% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractTech4}
\alias{extractTech4}
\title{Extract Technical 4 matrix from Mplus}
\usage{
extractTech4(outfiletext, filename)
}
\arguments{
\item{outfiletext}{the text of the output file}

\item{filename}{The name of the file}
}
\value{
A list of class \dQuote{mplus.tech4}
}
\description{
Function that extracts the Tech4 matrix
}
\examples{
# make me!!!
}
\seealso{
\code{\link{matrixExtract}}
}
\keyword{internal}

