% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractTech8}
\alias{extractTech8}
\title{Extract Technical 8 from Mplus}
\usage{
extractTech8(outfiletext, filename)
}
\arguments{
\item{outfiletext}{the text of the output file}

\item{filename}{The name of the file}
}
\value{
A list of class \dQuote{mplus.tech8}
}
\description{
The TECH8 option is used to print the optimization history of a model.
It also prints the potential scale reduction in Bayesian models.
}
\examples{
# make me!!!
}
\seealso{
\code{\link{matrixExtract}}
}
\keyword{internal}
