% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractValue}
\alias{extractValue}
\title{Extract values from Mplus output
An internal function used by extractSummaries_1file to extract
parameters from the output file using regular expressions.}
\usage{
extractValue(pattern, textToScan, filename, type = "int")
}
\arguments{
\item{pattern}{the exact text to be matched in the outfile that identifies the parameter of interest}

\item{textToScan}{the chunk of Mplus output to be parsed, passed as a vector of character strings (from the scan command).}

\item{filename}{the name of the file containing textToScan. Used to make more intelligible warning messages.}

\item{type}{the data type of the parameter, which determines the regexp used. Currently can be \dQuote{int}, \dQuote{dec}, \dQuote{str}, or \dQuote{calc}. Defaults to \dQuote{int}.}
}
\value{
A string or numeric vector
}
\description{
Extract values from Mplus output
An internal function used by extractSummaries_1file to extract
parameters from the output file using regular expressions.
}
\examples{
#make me!!!
}
\keyword{internal}
