% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractWarningsErrors_1file}
\alias{extractWarningsErrors_1file}
\title{Extract warnings and errors from 1 mplus file}
\usage{
extractWarningsErrors_1file(outfiletext, filename, input)
}
\arguments{
\item{outfiletext}{The text of the output file}

\item{filename}{The filename}

\item{input}{The input}
}
\value{
A list with two elements
  \item{errors}{Mplus Errors}
  \item{warnings}{Mplus Warnings}
}
\description{
Helper function
}
\examples{
# make me!!!
}
\keyword{internal}

