% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{filter_inp_filelist}
\alias{filter_inp_filelist}
\title{helper function to filter a set of Mplus input files based on whether the corresponding .out file already exists}
\usage{
filter_inp_filelist(inp_files, replaceOutfile = "always", quiet = TRUE)
}
\arguments{
\item{inp_files}{a vector of input file locations to check for corresponding .out files}

\item{replaceOutfile}{optional. Currently supports three settings: \dQuote{always}, which
runs all models, regardless of whether an output file for the model exists; \dQuote{never},
which does not run any model that has an existing output file; and \dQuote{modifiedDate}, which
only runs a model if the modified date for the input file is more recent than
the output file modified date (implying there have been updates to the model).}

\item{quiet}{whether to print out text indicating what files were skipped}
}
\description{
helper function to filter a set of Mplus input files based on whether the corresponding .out file already exists
}
\keyword{internal}
