% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{finalizeInitCollection}
\alias{finalizeInitCollection}
\title{Finalize Init Collection}
\usage{
finalizeInitCollection(initCollection)
}
\arguments{
\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Finalized initCollection
}
\description{
this function should handle initTags that still contain tags
once the initCollection is finalized, then process the deferred body tags
the notion is that the substitutions will be handled in an inefficient manner -- using lots
of regular expression parsing, not using the matched tags data.frame
}
\details{
we only need to handle simple and array tags
iterators should always be integers
foreach and conditional are not relevant

iterate over init tags until no tags are left
here, the init collection should already have had most of its tags substituted by
replaceInitTags above.
}
\keyword{internal}

