% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{friendlyGregexpr}
\alias{friendlyGregexpr}
\title{Friendly Regular Expression}
\usage{
friendlyGregexpr(pattern, charvector, perl = TRUE, include_tag = TRUE)
}
\arguments{
\item{pattern}{The pattern to search for}

\item{charvector}{Character vector}

\item{perl}{A logical whether or not to use perl based
regular expressions.  Defaults to \code{TRUE}.}

\item{include_tag}{if TRUE, return the match as a character string. This is the default, but
setting to \code{FALSE} is a bit faster}
}
\value{
A \code{data.frame}
}
\description{
Creates data frame documenting the start and end of all tags.
}
\examples{
\dontrun{
  friendlyGregexpr("(BY|WITH|ON)", 
    c("POS_WI BY", "X WITH Y WITH Z")
  )
} 
}
\author{
Michael Hallquist
}
\keyword{internal}
