% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{get_efa_nfactors}
\alias{get_efa_nfactors}
\title{Small helper function to obtain number of factors for an EFA output section}
\usage{
get_efa_nfactors(headers)
}
\arguments{
\item{headers}{a vector of EFA section headers from which to parse number of factors}
}
\value{
A vector of the number of factors in each heading, or a matrix if multilevel output
}
\description{
Small helper function to obtain number of factors for an EFA output section
}
\keyword{internal}
