% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{getInitTags}
\alias{getInitTags}
\title{Get Initial Tags}
\usage{
getInitTags(initCollection)
}
\arguments{
\item{initCollection}{A list?}
}
\value{
The initMatches
}
\description{
An internal function
}
\keyword{internal}
