% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getOutFileList}
\alias{getOutFileList}
\title{Get Output File List}
\usage{
getOutFileList(target, recursive = FALSE, filefilter)
}
\arguments{
\item{target}{The target file or directory}

\item{recursive}{A logical value whether to search recursively.
Defaults to \code{FALSE}.}

\item{filefilter}{A regular expression passed to \code{grep}
used to filter the output files.}
}
\value{
A character vector of the output files
}
\description{
This is a helper function used by extractModelSummaries and extractModelParameters.
It determines whether the target is a single file or a directory.
If it is a directory, all .out files are returned (perhaps recursively)
It also permits the files to be filtered using a certain regular expression.
}
\examples{
# make me!!!
}
\keyword{internal}
