% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getSection}
\alias{getSection}
\title{Get an Output Section}
\usage{
getSection(sectionHeader, outfiletext, headers = "standard")
}
\arguments{
\item{sectionHeader}{Header section}

\item{outfiletext}{Output file text}

\item{headers}{Can pass custom headers but defaults to a standard set.}
}
\value{
Section
}
\description{
Get an Output Section
}
\examples{
# make me!!!
}
\keyword{internal}
