% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphModels.R
\name{graphModel}
\alias{graphModel}
\title{Create a graphic model from Mplus}
\usage{
graphModel(model)
}
\arguments{
\item{The}{model}
}
\value{
a dot model
}
\description{
To do: add details
}
\keyword{internal}
