% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{htmlout}
\alias{htmlout}
\title{Read in Mplus Output File from HTML on the Mplus Website}
\usage{
htmlout(url)
}
\arguments{
\item{url}{The url giving an Mplus output file in HTML format.}
}
\value{
A character string that is the filename where the output file was written.
  Always a temp file in a temp directory.
}
\description{
This is an internal utility function to help with testing MplusAutomation.
On the Mplus website, many output files are included, which are good tests.
However, they are stored in HTML files. This function reads them in,
strips the HTML parts leaving just the output file, and then writes it to
disk and returns the file name, which can be used with \code{readModels()}
for testing.
}
\examples{

MplusAutomation:::htmlout("https://statmodel.com/usersguide/chap3/ex3.1.html")
}
\author{
Joshua F. Wiley <jwiley.psych@gmail.com>
}
\keyword{internal}
