% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcademo}
\alias{lcademo}
\title{Latent Class Analysis Demonstration}
\format{
A list containing 11 elements.
\itemize{
\item \code{Data}: 2 Class LCA data simulated using Mplus
\item \code{CFA}: Mplus output file for CFA
\item \code{LCA2}: Mplus output file for 2 class LCA
\item \code{LCA3}: Mplus output file for 3 class LCA
\item \code{LCA4}: Mplus output file for 4 class LCA
\item \code{LCA5}: Mplus output file for 5 class LCA
\item \code{CFAGH5}: GH5 file for CFA
\item \code{LCA2GH5}: GH5 file for 2 class LCA
\item \code{LCA3GH5}: GH5 file for 3 class LCA
\item \code{LCA4GH5}: GH5 file for 4 class LCA
\item \code{LCA5GH5}: GH5 file for 5 class LCA
}
}
\usage{
lcademo
}
\description{
A list containing the Mplus Data, Output Files, and GH5 Files for a
demonstration of using MplusAutomation for latent class analysis.
Generated by the vignette on latent class analysis.
}
\keyword{datasets}
