% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{local_job_status}
\alias{local_job_status}
\title{helper function to get job status on local *nix-based machine}
\usage{
local_job_status(
  job_ids = NULL,
  user = NULL,
  ps_format = "user,pid,state,time,etime,\%cpu,\%mem,comm,xstat"
)
}
\arguments{
\item{job_ids}{a vector of job ids (process IDs) to check}

\item{user}{the user who owns the processes (defaults to current user)}

\item{ps_format}{the format string passed to ps}
}
\value{
a data.table with job information for each job id
}
\description{
helper function to get job status on local *nix-based machine
}
\keyword{internal}
