% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainlgmm.R
\name{long2LGMM}
\alias{long2LGMM}
\title{Long data two wide latent growth mixture model}
\usage{
long2LGMM(data, idvar, assessmentvar, dv, timevars, misstrick = TRUE,
  k = 1L, title = "Trajectory Model", base = "trajmodel_", run = FALSE,
  processors = 1L, starts = "500 100", newdata, cov = c("un",
  "independent", "intercept", "zero"), model)
}
\arguments{
\item{data}{A data frame in long format (i.e., multiple rows per ID).}

\item{idvar}{A character string of the variable name in the dataset
that is the ID variable.}

\item{assessmentvar}{A character string of the variable name in the dataset that
indicates the particular assessment point for each timepoint.}

\item{dv}{A character string of the dependent variable name.}

\item{timevars}{A character vector of the time variables.
Can be a single variable or more than one.  By allowing more than one variable,
it is easy to include linear; linear and quadratic; it is also possible to
calculate splines in R and pass these. The variable names should be 7 characters
or fewer, each.}

\item{misstrick}{A logical value whether to set values of the DV where a time variable
is missing to missing as well.  Defaults to \code{TRUE}.}

\item{k}{An integer indicating the number of distinct classes to test.
Currently must be greater than 0 and less than 10.}

\item{title}{A character string giving a title for the model}

\item{base}{A character string providing a base name for model outputs,
that is combined with the number of classes.}

\item{run}{A logical value whether or not to run the models or only
create the data and input files, but not run them.}

\item{processors}{An integer value indicating the number of processors to use.}

\item{starts}{A character string passed to Mplus providing the number of
random starts and iterations}

\item{newdata}{A data frame of new values to use for generating predicted
trajectories by class.}

\item{cov}{A character string indicating the random covariance structure to use}

\item{model}{An optional argument, can pass an existing model, the output from
mplusModeler().}
}
\description{
This function streamlines the process of converting long data into a format
that Mplus can use for latent growth mixture models in wide form.  It makes
use of continuous time scores, and these time scores must be supplied as
variables in the R dataset.  For the conversion to wide form, it is assumed
that although assessments may have happened in continuous time, a discrete
number of assessments (likely ismilar for all participants) were collected.
}
\details{
One valuable feature of this function is that it makes it possible to feed
any continuous time scores to Mplus for mixture modelling.  For example,
continuous linear time is straightforward, but so to are quadratic
time models or piecewise models.  Using facilities in R, spline models
are also comparatively easy to specify.
}
\examples{
## Make me!
}
