% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{minutes_to_dhms}
\alias{minutes_to_dhms}
\title{helper function to convert a number of minutes into a dhms-formatted string for submission}
\usage{
minutes_to_dhms(x)
}
\arguments{
\item{x}{numeric or character number of minutes}
}
\value{
the hms string representing this number of minutes
}
\description{
helper function to convert a number of minutes into a dhms-formatted string for submission
}
\keyword{internal}
