\name{MplusAutomation-package}
\Rdversion{1.1}
\alias{MplusAutomation-package}
\alias{MplusAutomation}
\docType{package}
\title{Automating Mplus Model Estimation and Interpretation}
\description{
The MplusAutomation package leverages the flexibility of the 
	R language to automate latent variable model estimation and interpretation using Mplus, 
	a powerful latent variable modeling program developed by Muthen and Muthen (www.statmodel.com).
	Specifically, MplusAutomation provides routines for creating related groups of models,
	running batches of models, and extracting and tabulating model parameters and fit statistics.}
\details{
\tabular{ll}{
Package: \tab MplusAutomation\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2010-02-16\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Michael Hallquist

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
