% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{OS}
\alias{OS}
\alias{is.windows}
\alias{is.macos}
\alias{is.linux}
\alias{os}
\title{Functions to identify the operating system}
\usage{
is.windows()

is.macos()

is.linux()

os()
}
\value{
A logical value for the is.* functions and a character string for os()
}
\description{
Functions to identify the operating system
}
\examples{

MplusAutomation:::is.windows()

MplusAutomation:::is.macos()

MplusAutomation:::is.linux()

MplusAutomation:::os()
}
\author{
Joshua F. Wiley <jwiley.psych@gmail.com>
}
\keyword{internal}
