% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{parseMplusSyntax}
\alias{parseMplusSyntax}
\title{Convert an Mplus syntax string into a parsed list}
\usage{
parseMplusSyntax(syntax, dropSectionNames = TRUE)
}
\arguments{
\item{syntax}{the character vector containing Mplus syntax}

\item{dropSectionNames}{Logical. If TRUE, section names will be dropped from the parsed list}
}
\value{
a parsed list of Mplus syntax
}
\description{
Convert an Mplus syntax string into a parsed list
}
\details{
This function converts an Mplus syntax string into a list parsed into sections such as TITLE and DATA
}
\keyword{internal}
