% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{plotLTA}
\alias{plotLTA}
\title{Plot latent transition model}
\usage{
plotLTA(mplusModel, node_stroke = 2, max_edge_width = 2,
  node_labels = "variable.class", x_labels = "variable")
}
\arguments{
\item{mplusModel}{A single Mplus model object, returned by . This
function additionally requires the model to be a mixture model with multiple
categorical latent variables.}

\item{node_stroke}{Integer. Base stroke thickness for nodes. Set to 
\code{NULL} to give each node the same stroke thickness.}

\item{max_edge_width}{Integer. Maximum width of edges.}

\item{node_labels}{Character vector, defaults to \code{"variable.class"}, 
which labels each node by the name of the variable, and the number of the 
class it represents. Set to \code{"class"} to display only class numbers, or
provide a named character vector where the names correspond to original class
labels, and the values correspond to their substitute values.}

\item{x_labels}{Character vector, defaults to \code{"variable"}, which labels
the x-axis with the names of the categorical latent variables. Set to
\code{NULL} to remove axis labels, or provide a named character vector where
the names correspond to original x-axis labels, and the values correspond to
their substitute values.}
}
\value{
An object of class 'ggplot'.
}
\description{
Plots latent transition probabilities and classification probabilities for
a single latent transition model (a model with multiple categorical latent
variables, regressed on one another). Stroke thickness of nodes represents
the proportion of cases most likely assigned to that class, with wider 
strokes representing greater probability. Edge thickness and transparency
represent the probability of making a particular transition (left to right),
with thicker/darker edges representing greater probability.
}
\examples{
#Make me!!!
}
\author{
Caspar J. van Lissa
}
\keyword{internal}
