% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{processInit}
\alias{processInit}
\title{Process the Init Section}
\usage{
processInit(initsection)
}
\arguments{
\item{initsection}{The list of all arguments parsed from the init section}
}
\value{
arglist
}
\description{
To do: add details.
}
\keyword{internal}

