\name{recurseReplace}
\alias{recurseReplace}
\title{Recursive replace}
\usage{
  recurseReplace(templateTags, initCollection,
    curiterator = 1L)
}
\arguments{
  \item{templateTags}{The template tags}

  \item{initCollection}{The list of all arguments parsed
  from the init section}

  \item{curiterator}{An integer that tracks of the depth of
  recursion through the iterators. Defaults to 1.}
}
\value{
  Does not look like it returns anything
}
\description{
  To do: fill in some details
}
\keyword{internal}

