% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{replaceInitTags}
\alias{replaceInitTags}
\title{Replace Init Tags}
\usage{
replaceInitTags(initTags, initCollection)
}
\arguments{
\item{initTags}{Init tags}

\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Returns updated initCollection
}
\description{
To do: fill in some details
}
\keyword{internal}

