% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{seconds_to_dhms}
\alias{seconds_to_dhms}
\title{helper function to convert a number of seconds into an hms-formatted string}
\usage{
seconds_to_dhms(x)
}
\arguments{
\item{x}{numeric number of seconds}
}
\value{
an hms string suitable for scheduler submission
}
\description{
helper function to convert a number of seconds into an hms-formatted string
}
\keyword{internal}
