% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{separateHyphens}
\alias{separateHyphens}
\title{Separate Hyphenated Variable Strings}
\usage{
separateHyphens(cmd)
}
\arguments{
\item{cmd}{A character string}
}
\value{
The character string if no hyphens, or a list of vectors if there are hyphens.
}
\description{
This code is a simplified form of \code{expandCmd} from the \pkg{lavaan} package.
It separates hyphenated variable strings into a list of vectors, while ignoring
hyphens that may be used in numbers.
}
\details{
Note that this is an internal function only.
}
\examples{

MplusAutomation:::separateHyphens("x1x4")
MplusAutomation:::separateHyphens("x1-x4")
MplusAutomation:::separateHyphens("x1-x4; x1*-1; v1-v3;")
}
\author{
Michael Hallquist revised by Joshua Wiley
}
\keyword{interface}
