% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{slurm_job_status}
\alias{slurm_job_status}
\title{check the status of jobs submitted on a SLURM cluster}
\usage{
slurm_job_status(
  job_ids = NULL,
  user = NULL,
  sacct_format = "jobid,submit,timelimit,start,end,state"
)
}
\arguments{
\item{job_ids}{a vector of job ids to check}

\item{user}{the user who submitted the jobs}

\item{sacct_format}{the format string passed to sacct for job status}
}
\value{
a data.frame containing the status of each job id
}
\description{
check the status of jobs submitted on a SLURM cluster
}
\details{
This function calls \code{sacct -j} to check the status of jobs on a slurm cluster
}
\keyword{internal}
