% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{splitFilePath}
\alias{splitFilePath}
\title{Split File and Path into Separate Parts}
\usage{
splitFilePath(abspath)
}
\arguments{
\item{abspath}{A character string of the file path}
}
\value{
A list with elements for the directory, filename,
  and absolute path.
}
\description{
This is a helper function to split path into path and filename.
Code adapted from R.utils filePath command.
}
\examples{
# make me!!!
}
\keyword{internal}
