% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{torque_job_status}
\alias{torque_job_status}
\title{helper function to get the status of jobs on a TORQUE cluster}
\usage{
torque_job_status(job_ids, user = NULL)
}
\arguments{
\item{job_ids}{a vector of job ids to check}

\item{user}{the user who submitted the jobs}
}
\value{
a data.frame with the job status for each id
}
\description{
helper function to get the status of jobs on a TORQUE cluster
}
\details{
Torque does not keep information about completed jobs available in qstat or qselect.
Thus, we need to log when a job is listed as queued, so that it 'going missing' is evidence of it being completed.
}
\keyword{internal}
