% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{trimSpace}
\alias{trimSpace}
\title{Trim White Space}
\usage{
trimSpace(string)
}
\arguments{
\item{string}{The character vector to trim white space from.}
}
\value{
A character vector with the white space removed.
}
\description{
Helper function to remove white space from a character vector
}
\examples{
MplusAutomation:::trimSpace(c("    test", "another    "))
}
\keyword{internal}
