% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{validate_dhms}
\alias{validate_dhms}
\title{helper function to validate format of walltime inputs for HPC submission}
\usage{
validate_dhms(str)
}
\arguments{
\item{str}{string containing a duration that may include a days specification}
}
\description{
helper function to validate format of walltime inputs for HPC submission
}
\details{
this normalizes overflow in clock components (for example, 90
seconds becomes 1 minute and 30 seconds), converts to an hms format, and
retains a dhms format when days are present. Supported date formats match slurm sbatch:
https://slurm.schedmd.com/sbatch.html
}
\keyword{internal}
