\name{addHeaderToSavedata}
\alias{addHeaderToSavedata}
\title{Add header to saved data}
\usage{
  addHeaderToSavedata(outfile, directory = getwd())
}
\arguments{
  \item{outfile}{The output file}

  \item{director}{The current working directory by default}
}
\value{
  NULL
}
\description{
  Description
}
\examples{
# make me!!!
}
\keyword{internal}

