% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clipString}
\alias{clipString}
\title{Clip String}
\usage{
clipString(string, start, end)
}
\arguments{
\item{string}{A string to be clipped}

\item{start}{The character position to start at}

\item{end}{The character position to end at}
}
\value{
A string from start to end
}
\description{
To do: add any details.
}
\keyword{internal}

