% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphModels.R
\name{connectNodes}
\alias{connectNodes}
\title{Connect two nodes}
\usage{
connectNodes(dotModel, node1, node2, connectionType)
}
\arguments{
\item{dotModel}{The basic model}

\item{node1}{The starting node}

\item{node2}{The ending node}

\item{connectionType}{The type of connection to add between nodes 1 and 2}
}
\description{
To do: add details
}
\keyword{internal}

