\name{createTable}
\Rdversion{1.1}
\alias{createTable}

\title{Create HTML/LaTex Table from parsed Mplus Models}   
\description{
Creates an HTML or LaTex table that displays model summary statistics for a variety of parsed Mplus models.
This function expects to receive a \code{data.frame} generated by the \code{extractModelSummaries} command. If \code{latex}
is \code{FALSE}, then the output is an HTML file specified by \code{filename}. Otherwise, a LaTex table generated
by \code{xtable} is returned.
}
\usage{
createTable(modelList, filename = file.path(getwd(), "Model Comparison.html"),
	sortby = "AICC", display = TRUE, latex = FALSE, 
	dropCols = c("InputInstructions", "Observations"), label = NULL)
}
\arguments{
  \item{modelList}{A list of models (as a \code{data.frame}) returned from the \code{extractModelSummaries}
  	function.}
  \item{filename}{The name of the HTML file to be created. Can be an absolute or relative path.
  	If \code{filename} is a relative path or just the filename, then it is assumed that the file resides in
  	the working directory \code{getwd()}.}
  \item{sortby}{optional. Field name (as character string) by which to sort the table.
  	Typically an information criterion (e.g., "AIC" or "BIC") is used to sort the table.}
  \item{display}{optional. When creating an HTML file containing the table, this parameter
  	specifies whether to display the table upon creation (\code{TRUE} or \code{FALSE}).}
  \item{latex}{optional. The \code{createTable} function will return a LaTex table
  	if \code{latex} is \code{TRUE}. If \code{FALSE}, the function will write the table to
  	an HTML file specified by \code{filename}.}
  \item{dropCols}{optional. A vector of character strings specifying which columns should be
  	dropped from the \code{modelList} data frame prior to creating the table. Defaults to
  	c("inputInstructions", "Observations").}
  \item{label}{optional. A character string specifying the LaTex label for the table.
  	Only relevant when \code{latex} is \code{TRUE}.}
}
\value{
If \code{latex} is \code{TRUE}, then a LaTex-formatted table summarizing the \code{modelList}
is returned (created by \code{xtable}). Otherwise, nothing is returned because an HTML file is created.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{Michael Hallquist}

\seealso{
\code{\link{extractModelSummaries}}

\code{\link{xtable}}
}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{interface}