% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAuxDis.R
\name{extractAux}
\alias{extractAux}
\title{Extract Auxiliary condition means and comparisons.}
\usage{
extractAux(outfiletext, filename)
}
\arguments{
\item{outfiletext}{character vector of Mplus output file from which to extract the AUX section}

\item{filename}{filename of the Mplus output file being processed}
}
\value{
A data frame
}
\description{
To do: add details.
}
\keyword{internal}
