% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{extractClassCounts}
\alias{extractClassCounts}
\title{Extract Latent Class Counts}
\usage{
extractClassCounts(outfiletext, filename, summaries)
}
\arguments{
\item{outfiletext}{The text of the output file}

\item{filename}{The name of the output file}
}
\value{
a list
}
\description{
Function for extracting counts of latent classes
}
\examples{
# make me!!!
}
\keyword{internal}
