% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndirect.R
\name{extractIndirect}
\alias{extractIndirect}
\title{Extract Indirect Effects output}
\usage{
extractIndirect(outfiletext, curfile)
}
\arguments{
\item{outfiletext}{a character vector containing the indirect effects output section returned by getSection}

\item{curfile}{the name of the current output file being parsed}
}
\value{
An mplus.indirect object (inherits list) containing $overall and $specific
}
\description{
This function parses both unstandardized and standardized indirect effects
It returns a list composed of $unstandardized and $standardized.
The base structure of each is a list containing $overall and $specific effects (as data.frames)
}
\keyword{internal}
