\name{extractModIndices_1file}
\alias{extractModIndices_1file}
\title{Extract Modification Indices for One File}
\usage{
  extractModIndices_1file(outfiletext, filename)
}
\arguments{
  \item{outfiletext}{}

  \item{filename}{}
}
\value{
  A data frame
}
\description{
  To do: add details.
}
\keyword{internal}

