% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractModIndices_1file}
\alias{extractModIndices_1file}
\title{Extract Modification Indices for One File}
\usage{
extractModIndices_1file(outfiletext, filename)
}
\arguments{
\item{outfiletext}{}

\item{filename}{}
}
\value{
A data frame
}
\description{
To do: add details.
}
\keyword{internal}

