% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParameters.R
\name{extractParameters_1file}
\alias{extractParameters_1file}
\title{Extract Parameters for One File}
\usage{
extractParameters_1file(outfiletext, filename, resultType)
}
\arguments{
\item{outfiletext}{character vector of Mplus output file being processed}

\item{filename}{name of Mplus output file being processed}

\item{resultType}{(deprecated)}
}
\value{
A list of parameters
}
\description{
To do: add details
}
\keyword{internal}

