\name{extractParameters_1file}
\alias{extractParameters_1file}
\title{Extract Parameters for One File}
\usage{
  extractParameters_1file(outfiletext, filename,
    resultType)
}
\arguments{
  \item{filename}{}

  \item{modelSection}{}

  \item{sectionName}{}
}
\value{
  A list of parameters
}
\description{
  To do: add details
}
\keyword{internal}

