% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{friendlyGregexpr}
\alias{friendlyGregexpr}
\title{Friendly Regular Expression}
\usage{
friendlyGregexpr(pattern, charvector, perl = TRUE)
}
\arguments{
\item{pattern}{The pattern to search for}

\item{charvector}{Character vector}

\item{perl}{A logical whether or not to use perl based
regular expressions.  Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame}
}
\description{
Creates data frame documenting the start and end of all tags.
}
\examples{
## make me
}
\author{
Michael Hallquist
}
\keyword{internal}
