% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getMultilineSection}
\alias{getMultilineSection}
\title{Extract a multiline section from Mplus output}
\usage{
getMultilineSection(
  header,
  outfiletext,
  filename,
  allowMultiple = FALSE,
  allowSpace = TRUE,
  ignore.case = FALSE
)
}
\arguments{
\item{header}{Header section}

\item{outfiletext}{Output file text}

\item{filename}{The name of the file}

\item{allowMultiple}{Logical indicating whether to allow multiple sections. Defaults to \code{FALSE}.}

\item{allowSpace}{Logical indicating whether to allow spaces. Defaults to \code{TRUE}.}

\item{ignore.case}{Logical whether or not to ignore the case.  Defaults to \code{FALSE}.}
}
\value{
A list of sections
}
\description{
New approach to multiline section: retain spaces and look for next line that has identical indentation.
}
\examples{
# make me!!!
}
\keyword{internal}
