% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{lookupValue}
\alias{lookupValue}
\title{Lookup values}
\usage{
lookupValue(tag, tagType, initCollection)
}
\arguments{
\item{tag}{name of tag for which we want to know the current value}

\item{tagType}{type of tag (simple, array, etc.) for the tag to lookup}

\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Current value
}
\description{
To do: fill in some details
}
\keyword{internal}
