% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lookupValue}
\alias{lookupValue}
\title{Lookup values}
\usage{
lookupValue(tag, tagType, initCollection)
}
\arguments{
\item{tag}{}

\item{tagType}{}

\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Current value
}
\description{
To do: fill in some details
}
\keyword{internal}

