% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrixExtract}
\alias{matrixExtract}
\title{Reconstruct matrix from multi-line text output}
\usage{
matrixExtract(outfiletext, headerLine, filename)
}
\arguments{
\item{outfiletext}{The text of the output file}

\item{headerLine}{The header line}

\item{filename}{The name of the output file}
}
\value{
a matrix
}
\description{
main worker function for extracting Mplus matrix output
where matrices are spread across blocks to keep within width constraints
example: tech1 matrix output.
}
\examples{
# make me!!!
}
\keyword{internal}

