% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{matrixExtract}
\alias{matrixExtract}
\title{Reconstruct matrix from multi-line text output}
\usage{
matrixExtract(
  outfiletext,
  headerLine,
  filename,
  ignore.case = FALSE,
  expect_sig = FALSE
)
}
\arguments{
\item{outfiletext}{The text of the output file}

\item{headerLine}{The header line}

\item{filename}{The name of the output file}

\item{ignore.case}{Whether to ignore case of header line}

\item{expect_sig}{Whether to track value significance TRUE/FALSE (* in Mplus) as an attribute}
}
\value{
a matrix
}
\description{
main worker function for extracting Mplus matrix output
where matrices are spread across blocks to keep within width constraints
example: tech1 matrix output.
}
\examples{
# make me!!!
}
\keyword{internal}
