% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{minutes_to_dhms}
\alias{minutes_to_dhms}
\title{helper function to convert a number of minutes into a dhms-formatted string for submission}
\usage{
minutes_to_dhms(x)
}
\arguments{
\item{x}{string or charcater number of minutes}
}
\value{
the dhms string representing this number of minutes in days, hours, minutes, and seconds
}
\description{
helper function to convert a number of minutes into a dhms-formatted string for submission
}
\keyword{internal}
