% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplusModel.R
\name{mplusModel}
\alias{mplusModel}
\title{Create an mplusModel object for a given model}
\usage{
mplusModel(
  syntax = NULL,
  data = NULL,
  inp_file = NULL,
  read = TRUE,
  Mplus_command = NULL
)
}
\arguments{
\item{syntax}{a character vector of Mplus input syntax for this model}

\item{data}{a data.frame to be used for estimating the model}

\item{inp_file}{the location of .inp file for this model}

\item{read}{If TRUE and the .out file already exists, read the contents of the .out file using \code{readModels}}

\item{Mplus_command}{The location of the Mplus executable to run. If NULL, use \code{detectMplus()}}
}
\value{
a \code{mplusModel_r6} object containing information about the model
}
\description{
Create an mplusModel object for a given model
}
