% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplusModel.R
\name{mplusModel_r6}
\alias{mplusModel_r6}
\title{An R6 class for a single Mplus model}
\description{
An R6 class for a single Mplus model

An R6 class for a single Mplus model
}
\details{
Wrapped by `mplusModel`
  Note that this R6 class deliberately uses unlockBinding to populate private fields after the object is instantiated.
  This allows the model outputs to be added to the relevant sections of the object while keeping the fields private.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{input}}{Mplus input syntax parsed into a list by major section}

\item{\code{warnings}}{Syntax and estimation warnings as a list}

\item{\code{parameters}}{a list containing the parameter estimates for the model}

\item{\code{summaries}}{a list containing the model summary information and statistics}

\item{\code{savedata}}{a list containing the data output by the SAVEDATA command}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model_dir}}{the directory for Mplus files corresponding to this model}

\item{\code{inp_file}}{the location of the Mplus .inp file for this model}

\item{\code{out_file}}{the location of the Mplus .out file for this model}

\item{\code{dat_file}}{the location of the Mplus .dat (data) file for this model}

\item{\code{data}}{the dataset used for estimating this model}

\item{\code{Mplus_command}}{the location of the Mplus program}

\item{\code{syntax}}{the Mplus syntax for this model as a character vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mplusModel_r6-new}{\code{mplusModel_r6$new()}}
\item \href{#method-mplusModel_r6-read}{\code{mplusModel_r6$read()}}
\item \href{#method-mplusModel_r6-write_dat}{\code{mplusModel_r6$write_dat()}}
\item \href{#method-mplusModel_r6-write_inp}{\code{mplusModel_r6$write_inp()}}
\item \href{#method-mplusModel_r6-submit}{\code{mplusModel_r6$submit()}}
\item \href{#method-mplusModel_r6-run}{\code{mplusModel_r6$run()}}
\item \href{#method-mplusModel_r6-clone}{\code{mplusModel_r6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-new"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-new}{}}}
\subsection{Method \code{new()}}{
generate an mplusModel_r6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$new(
  syntax = NULL,
  data = NULL,
  inp_file = NULL,
  read = TRUE,
  Mplus_command = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{syntax}}{a character vector of Mplus input syntax for this model}

\item{\code{data}}{a data.frame to be used for estimating the model}

\item{\code{inp_file}}{the location of .inp file for this model}

\item{\code{read}}{If TRUE and the .out file already exists, read the contents of the .out file using `readModels`}

\item{\code{Mplus_command}}{N.B.: No need to pass this parameter for most users (has intelligent defaults). Allows the user to 
specify the name/path of the Mplus executable to be used for running models. This covers situations where Mplus
is not in the system's path, or where one wants to test different versions of the Mplus program.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-read"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-read}{}}}
\subsection{Method \code{read()}}{
read the results of an Mplus model from the .out file using `readModels`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$read(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{if `TRUE`, re-read the .out file even if was already previously loaded in to this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-write_dat"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-write_dat}{}}}
\subsection{Method \code{write_dat()}}{
write the .inp and .dat files for this model to the intended location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$write_dat(overwrite = TRUE, quiet = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{overwrite}}{if `TRUE`, overwrite existing data. Default: `TRUE`.}

\item{\code{quiet}}{if `TRUE`, do not produce messages about the outcome of this command (e.g., a message about overwriting existing data)}

\item{\code{...}}{additional arguments passed to `prepareMplusData`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-write_inp"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-write_inp}{}}}
\subsection{Method \code{write_inp()}}{
write the .inp and .dat files for this model to the intended location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$write_inp(overwrite = TRUE, inp_file = NULL, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{overwrite}}{if `TRUE`, overwrite existing data. Default: `TRUE`.}

\item{\code{inp_file}}{The location of the input file to write. If NULL (default), use the `$inp_file` of this object.}

\item{\code{quiet}}{if `TRUE`, do not produce messages about the outcome of this command (e.g., a message about overwriting existing data)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-submit"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-submit}{}}}
\subsection{Method \code{submit()}}{
submit this model for estimation on an HPC using `submitModels`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$submit(replaceOutfile = "modifiedDate", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replaceOutfile}}{Currently supports three settings: “always”, which runs all models, regardless of whether an output file for the model exists; 
“never”, which does not run any model that has an existing output file; and “modifiedDate”, which only runs a model if the modified date for the input
file is more recent than the output file modified date (implying there have been updates to the model).}

\item{\code{...}}{additional arguments passed to `submitModels`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-run"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-run}{}}}
\subsection{Method \code{run()}}{
run this model locally using `runModels`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$run(replaceOutfile = "modifiedDate", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replaceOutfile}}{Currently supports three settings: “always”, which runs all models, regardless of whether an output file for the model exists; 
“never”, which does not run any model that has an existing output file; and “modifiedDate”, which only runs a model if the modified date for the input
file is more recent than the output file modified date (implying there have been updates to the model).}

\item{\code{...}}{additional arguments passed to `runModels`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mplusModel_r6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mplusModel_r6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mplusModel_r6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
