% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutput.R
\name{parseMplusSyntax}
\alias{parseMplusSyntax}
\title{Convert an Mplus syntax string into a parsed list}
\usage{
parseMplusSyntax(syntax, dropSectionNames = TRUE)
}
\description{
Convert an Mplus syntax string into a parsed list
}
