% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{processConditionalTags}
\alias{processConditionalTags}
\title{Process Conditional Tags}
\usage{
processConditionalTags(templateTags, initCollection)
}
\arguments{
\item{templateTags}{A template tag}

\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Processed templateTags
}
\description{
To do: add details.
}
\keyword{internal}
