% Generated by roxygen2 (4.0.1): do not edit by hand
\name{processInit}
\alias{processInit}
\title{Process the Init Section}
\usage{
processInit(initsection)
}
\arguments{
\item{initsection}{The list of all arguments parsed from the init section}
}
\value{
arglist
}
\description{
To do: add details.
}
\keyword{internal}

