% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{replaceBodyTags}
\alias{replaceBodyTags}
\title{Replace Body Tags}
\usage{
replaceBodyTags(bodySection, bodyTags, initCollection)
}
\arguments{
\item{bodySection}{character vector of body section of Mplus syntax}

\item{bodyTags}{collection of tags used inside of the template body}

\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Returns updated bodySection
}
\description{
To do: fill in some details
}
\keyword{internal}

