% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replaceBodyTags}
\alias{replaceBodyTags}
\title{Replace Body Tags}
\usage{
replaceBodyTags(bodySection, bodyTags, initCollection)
}
\arguments{
\item{bodySection}{}

\item{bodyTags}{}

\item{initCollection}{The list of all arguments parsed from the init section}
}
\value{
Returns updated bodySection
}
\description{
To do: fill in some details
}
\keyword{internal}

