% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitModels.R
\name{summary.mplus_submission_df}
\alias{summary.mplus_submission_df}
\title{summary function for submission from \code{submitModels}}
\usage{
\method{summary}{mplus_submission_df}(object, refresh = TRUE, ...)
}
\arguments{
\item{object}{the \code{mplus_submission_df} object to summarize}

\item{refresh}{if \code{TRUE}, check the status of jobs for this object before printing}

\item{...}{additional arguments, not currently used}
}
\description{
summary function for submission from \code{submitModels}
}
